requirements.txt
Django==5.0.6
django-extensions==3.2.3
djangorestframework==3.15.1
drf-nested-routers==0.94.1
drf-spectacular==0.27.2
factory-boy==3.3.0
Faker==25.2.0
pydantic==2.7.1
pydantic_core==2.18.2
requests==2.31.0
